Remove test_quantiles to cudf directories
