+++ python/cudf/cudf/tests/dataframe/methods/test_reductions.py
@@ -1,10 +1,11 @@
 # Copyright (c) 2025, NVIDIA CORPORATION.
 
-
 import numpy as np
+import pandas as pd
 import pytest
 
 import cudf
+from cudf.testing import assert_eq
 
 
 @pytest.mark.parametrize("null_flag", [False, True])
@@ -60,3 +61,47 @@ def test_skew_df(null_flag, numeric_only):
     expected = pdata.skew(numeric_only=numeric_only)
     got = got if np.isscalar(got) else got.to_numpy()
     np.testing.assert_array_almost_equal(got, expected)
+
+
+def test_single_q():
+    q = 0.5
+
+    pdf = pd.DataFrame({"a": [4, 24, 13, 8, 7]})
+    gdf = cudf.from_pandas(pdf)
+
+    pdf_q = pdf.quantile(q, interpolation="nearest")
+    gdf_q = gdf.quantile(q, interpolation="nearest", method="table")
+
+    assert_eq(pdf_q, gdf_q, check_index_type=False)
+
+
+def test_with_index():
+    q = [0, 0.5, 1]
+
+    pdf = pd.DataFrame({"a": [7, 4, 4, 9, 13]}, index=[0, 4, 3, 2, 7])
+    gdf = cudf.from_pandas(pdf)
+
+    pdf_q = pdf.quantile(q, interpolation="nearest")
+    gdf_q = gdf.quantile(q, interpolation="nearest", method="table")
+
+    assert_eq(pdf_q, gdf_q, check_index_type=False)
+
+
+def test_with_multiindex():
+    q = [0, 0.5, 1]
+
+    pdf = pd.DataFrame(
+        {
+            "index_1": [3, 1, 9, 7, 5],
+            "index_2": [2, 4, 3, 5, 1],
+            "a": [8, 4, 2, 3, 8],
+        }
+    )
+    pdf.set_index(["index_1", "index_2"], inplace=True)
+
+    gdf = cudf.from_pandas(pdf)
+
+    pdf_q = pdf.quantile(q, interpolation="nearest")
+    gdf_q = gdf.quantile(q, interpolation="nearest", method="table")
+
+    assert_eq(pdf_q, gdf_q, check_index_type=False)
